requirements.txt
pyld
cryptography
py-multibase
jwcrypto
base58